+++ semcore/time-picker/src/TimePicker.jsx
@@ -40,7 +40,7 @@ export function formatHoursTo12(hours /* hours by 24 */) {
 
   // if not (HH:00)
   if (nHours === 0) return 12; // 0 => 12 PM
-  if (nHours > 12) return nHours - 12; // 22 => 12 PM
+  if (nHours > 12) return nHours - 12; // 22 => 10 PM
 
   return hours;
 }
@@ -78,6 +78,7 @@ class TimePickerRoot extends Component {
     ),
     i18n: localizedMessages,
     locale: 'en',
+    defaultTitle: '',
   });
 
   hoursInputRef = React.createRef();
@@ -88,7 +89,19 @@ class TimePickerRoot extends Component {
   uncontrolledProps() {
     return {
       value: null,
+      title: null,
     };
+  }
+
+  componentDidMount() {
+    const { id, 'aria-describedby': ariaDescribedBy } = this.asProps;
+    const selector = `[for=${id}]`;
+    const titleElement =
+      document.querySelector(selector) ?? document.querySelector(`#${ariaDescribedBy}`);
+
+    if (titleElement) {
+      this.handlers.title(titleElement.textContent);
+    }
   }
 
   get value() {
@@ -195,14 +208,12 @@ class TimePickerRoot extends Component {
   }
 
   getFormatProps() {
-    const { size, disabled, disablePortal, value, getI18nText } = this.asProps;
-    const valueFulfilled = value?.split(':').every((chunk) => chunk.length > 0);
+    const { size, disabled, disablePortal, getI18nText } = this.asProps;
 
     return {
       size,
       disabled,
       disablePortal,
-      ['aria-hidden']: !valueFulfilled,
       meridiem: this.meridiem,
       onClick: this.handleMeridiemClick,
       getI18nText,
@@ -211,15 +222,27 @@ class TimePickerRoot extends Component {
 
   render() {
     const STimePicker = Root;
-    const { styles, Children, value, is12Hour, getI18nText } = this.asProps;
+    const { styles, Children, value, is12Hour, getI18nText, title } = this.asProps;
+    const [hours, minutes] = this.valueToTime(this.value);
+
     const label = value
-      ? getI18nText('title', { time: value, meridiem: is12Hour ? this.meridiem : '' })
-      : getI18nText('titleEmpty');
+      ? `${title} ${getI18nText('title', {
+          time: `${hours}:${withLeadingZero(minutes)}`,
+          meridiem: is12Hour ? this.meridiem : '',
+        })}`
+      : `${title} ${getI18nText('titleEmpty')}`;
 
     return sstyled(styles)(
-      <STimePicker render={Input} aria-label={label} aria-valuenow={value || undefined}>
-        <Children />
-      </STimePicker>,
+      <>
+        <STimePicker
+          render={Input}
+          role={'group'}
+          aria-label={label}
+          __excludeProps={['value', 'title']}
+        >
+          <Children />
+        </STimePicker>
+      </>,
     );
   }
 }