+++ semcore/feedback-form/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [6.33.2] - 2024-08-02
+
+### Fixed
+
+- View for close icon as a `Button` component.
+- React errors in console about `validationOnBlur` property.
+
 ## [6.33.1] - 2024-07-30
 
 ### Changed