+++ semcore/i18n-unplugin/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [0.18.3] - 2024-08-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/core` [2.29.1 ~> 2.29.2]).
+
 ## [0.18.2] - 2024-07-30
 
 ### Changed