+++ semcore/input-tags/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.36.2] - 2024-08-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.32.1 ~> 4.32.2], `@semcore/input` [4.32.1 ~> 4.32.2], `@semcore/scroll-area` [5.34.1 ~> 5.35.0], `@semcore/core` [2.29.1 ~> 2.29.2]).
+
 ## [4.36.1] - 2024-07-30
 
 ### Changed