+++ semcore/pagination/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.42.2] - 2024-08-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/button` [5.34.1 ~> 5.35.0], `@semcore/flex-box` [5.31.1 ~> 5.31.2], `@semcore/icon` [4.43.1 ~> 4.43.2], `@semcore/input` [4.32.1 ~> 4.32.2], `@semcore/link` [5.36.1 ~> 5.36.2], `@semcore/tooltip` [6.40.1 ~> 6.40.2], `@semcore/utils` [4.32.1 ~> 4.32.2], `@semcore/core` [2.29.1 ~> 2.29.2]).
+
 ## [4.42.1] - 2024-07-30
 
 ### Changed